requirements/test-requirements.txt
-r requirements.txt
pytest-django
model-mommy
